refactor xhr to HttpUtils and create a download function
