+++ src/lib/API.js
@@ -1,7 +1,7 @@
 import _ from 'underscore';
 import Ion from './Ion';
 import IONKEYS from '../IONKEYS';
-import xhr from './xhr';
+import HttpUtils from './HttpUtils';
 import NetworkConnection from './NetworkConnection';
 import CONFIG from '../CONFIG';
 import * as Pusher from './Pusher/pusher';
@@ -92,7 +92,7 @@ function createLogin(login, password) {
     // Using xhr instead of request becasue request has logic to re-try API commands when we get a 407 authToken expired
     // in the response, and we call CreateLogin after getting a successful resposne to Authenticate so it's unlikely
     // that we'll get a 407.
-    return xhr('CreateLogin', {
+    return HttpUtils.xhr('CreateLogin', {
         authToken,
         partnerName: CONFIG.EXPENSIFY.PARTNER_NAME,
         partnerPassword: CONFIG.EXPENSIFY.PARTNER_PASSWORD,
@@ -151,7 +151,7 @@ function request(command, parameters, type = 'post') {
     // with 407 authToken expired. When other api commands fail with this error we call Authenticate
     // to get a new authToken and then fire the original api command again
     if (command === 'Authenticate') {
-        return xhr(command, parameters, type)
+        return HttpUtils.xhr(command, parameters, type)
             .then((response) => {
                 // If we didn't get a 200 response from authenticate we either failed to authenticate with
                 // an expensify login or the login credentials we created after the initial authentication.
@@ -191,7 +191,7 @@ function request(command, parameters, type = 'post') {
 
     // Make the http request, and if we get 407 jsonCode in the response,
     // re-authenticate to get a fresh authToken and make the original http request again
-    return xhr(command, parametersWithAuthToken, type)
+    return HttpUtils.xhr(command, parametersWithAuthToken, type)
         .then((responseData) => {
             // We can end up here if we have queued up many
             // requests and have an expired authToken. In these cases,
@@ -204,7 +204,7 @@ function request(command, parameters, type = 'post') {
             // we re-authenticate and then re-try the original request
             if (responseData.jsonCode === 407 && parametersWithAuthToken.doNotRetry !== true) {
                 reauthenticating = true;
-                return xhr('Authenticate', {
+                return HttpUtils.xhr('Authenticate', {
                     useExpensifyLogin: false,
                     partnerName: CONFIG.EXPENSIFY.PARTNER_NAME,
                     partnerPassword: CONFIG.EXPENSIFY.PARTNER_PASSWORD,
@@ -228,7 +228,7 @@ function request(command, parameters, type = 'post') {
                         Ion.merge(IONKEYS.SESSION, _.pick(response, 'authToken'));
                         return response;
                     })
-                    .then(() => xhr(command, parametersWithAuthToken, type))
+                    .then(() => HttpUtils.xhr(command, parametersWithAuthToken, type))
                     .catch((error) => {
                         reauthenticating = false;
                         redirectToSignIn(error.message);
@@ -256,7 +256,7 @@ function processNetworkRequestQueue() {
         // 2. Getting a 200 response back from the API (happens right below)
 
         // Make a simple request every second to see if the API is online again
-        xhr('Get', {doNotRetry: true})
+        HttpUtils.xhr('Get', {doNotRetry: true})
             .then(() => NetworkConnection.setOfflineStatus(false));
         return;
     }
